Move CPU count function to common, fix wrong CPU count in visited list
