+++ lib/common/common/Cargo.toml
@@ -10,6 +10,7 @@ edition = "2021"
 publish = false
 
 [dependencies]
+num_cpus = "1.16"
 ordered-float = "4.2"
 serde = { workspace = true }
 validator = { version = "0.16", features = ["derive"] }
