+++ lib/common/common/src/cpu.rs
@@ -3,6 +3,22 @@ use thiserror::Error;
 #[cfg(target_os = "linux")]
 use thread_priority::{set_current_thread_priority, ThreadPriority, ThreadPriorityValue};
 
+/// Try to read number of CPUs from environment variable `QDRANT_NUM_CPUS`.
+/// If it is not set, use `num_cpus::get()`.
+pub fn get_num_cpus() -> usize {
+    match std::env::var("QDRANT_NUM_CPUS") {
+        Ok(val) => {
+            let num_cpus = val.parse::<usize>().unwrap_or(0);
+            if num_cpus > 0 {
+                num_cpus
+            } else {
+                num_cpus::get()
+            }
+        }
+        Err(_) => num_cpus::get(),
+    }
+}
+
 #[derive(Error, Debug)]
 #[cfg(target_os = "linux")]
 pub enum ThreadPriorityError {
