--- lib/segment/src/common/cpu.rs
@@ -1,15 +1,0 @@
-/// Try to read number of CPUs from environment variable `QDRANT_NUM_CPUS`.
-/// If it is not set, use `num_cpus::get()`.
-pub fn get_num_cpus() -> usize {
-    match std::env::var("QDRANT_NUM_CPUS") {
-        Ok(val) => {
-            let num_cpus = val.parse::<usize>().unwrap_or(0);
-            if num_cpus > 0 {
-                num_cpus
-            } else {
-                num_cpus::get()
-            }
-        }
-        Err(_) => num_cpus::get(),
-    }
-}
