+++ lib/segment/src/common/mod.rs
@@ -1,6 +1,5 @@
 pub mod anonymize;
 pub mod arc_atomic_ref_cell_iterator;
-pub mod cpu;
 pub mod error_logging;
 pub mod mmap_type;
 pub mod operation_error;
