+++ lib/storage/src/content_manager/toc/mod.rs
@@ -27,9 +27,9 @@ use collection::shards::replica_set;
 use collection::shards::replica_set::{AbortShardTransfer, ReplicaState};
 use collection::shards::shard::{PeerId, ShardId};
 use collection::telemetry::CollectionTelemetry;
+use common::cpu::get_num_cpus;
 use futures::future::try_join_all;
 use futures::Future;
-use segment::common::cpu::get_num_cpus;
 use tokio::runtime::Runtime;
 use tokio::sync::{Mutex, RwLock, RwLockReadGuard, Semaphore};
 use tonic::codegen::InterceptedService;
